♻️ :: Application 쪽 UseCase 하나로 합치기
